requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
yt-dlp==2023.12.30
mutagen==1.47.0
PyNaCl==1.5.0